+++ scripts/index.js
@@ -1,4 +1,6 @@
 #!/usr/bin/env node
+console.log('DEVELOPMENT');
+
 const cli = require('cli');
 
 cli.parse(null, ['dev', 'build', 'new', 'production']);