+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# 1.3.0 [2019-05-10]
+- you can extend webpack configs by adding
+  - `webpack.config.js` global changes
+  - `client.webpack.config.js` client webpack changes
+  - `server.webpack.config.js` server webpack changes
+  - `babel.config.js` babel changes
+
 # 1.2.0 [2019-05-10]
 - you can replace the default index.html by updating `src/index.html`